CBC-1840: Travel Sample using Spring Data Couchbase
This is the Travel Sample using Spring Data Couchbase. It is based
on the Java Travel Sample. All the services except for the Hotel
service (which uses FTS) have been converted to use Spring Data
Couchbase repositories.

The implementation relies on collections support in Spring Data
Couchbase which has not yet been released, therefore the pom
depdendency for spring data excludes spring-data-couchbase and
there is a separate dependency for 4.3.0-SNAPSHOT.  When
collections support in spring-data-couchbase has been released,
the separate dependency and the `exclude` can be removed.